requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
py-clob-client>=0.17.0
beautifulsoup4>=4.12.0
anthropic>=0.80.0
google-genai>=1.0.0
